Always echo cert password to openssl
This commit always passes the password for a PKCS#12 archive cert to openssl, to convert
to PEM format, even when the password is an empty string.

Use the insecure flag when calling localhost over HTTPS with curl. Since the subject name in certificate
used to secure the HTTP layer may not match host name localhost (it's likely to be tied to a public domain
name), the --cacert flag for curl cannot be used.